max_performance_rag/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0

pydantic==2.9.2

sentence-transformers==2.6.0

chromadb==0.5.23

streamlit==1.41.1
requests==2.32.5

pypdf==3.17.1
python-docx==1.1.0
python-dotenv==1.0.0

ollama==0.4.7
python-dotenv
slowapi